+++ lib/.byebug_history
@@ -1,0 +1,12 @@
+c
+cal.send :to_bs_date, bs_date
+bs_date = "2079-1-1"
+cal.send :to_bs_date, bs_date
+bs_date
+cal.to_bs_date bs_date
+cal = NepaliCalendar::BsCalendar.new "", year: 2079, month: 1, day: 1
+cal = NepaliCalendar::BsCalendar.new
+to_bs_date(bs_date)
+bs_date.to_bs_date
+bs_date.class
+bs_date
